requirements.txt
fastapi
uvicorn[standard]
google-cloud-secret-manager
google-cloud-storage
google-cloud-firestore
python-dotenv
cryptography # For in-memory decryption example